refactor: move calculatePosition method.
