+++ src/main/kotlin/common/token/Position.kt
@@ -10,16 +10,3 @@ data class Position(val row: Int, val column: Int) : Comparable<Position> {
     }
 }
 
-fun calculatePosition(code: String, index: Int): Position {
-    var row = 1
-    var column = 1
-    for (i in 0 until index) {
-        if (code[i] == '\n') {
-            row++
-            column = 1
-        } else {
-            column++
-        }
-    }
-    return Position(row, column)
-}
